Refactor: Convert CourseLeaderboardEntry to TypeScript
